update : coin model

+++ coin/models.py
@@ -4,6 +4,6 @@ from django.db import models
 class CoinTypes(models.Model):
     name = models.CharField(max_length=100)
     symbol = models.CharField(max_length=100)
-    price = models.FloatField(
+    price = models.IntegerField(
         default=0,
     )
